Another commit was made during the progress of the last commit I made, so this resolves the errors that were brought up.

+++ src/main/java/frc/robot/commands/Serializer/checkIfKickerUpToSpeed.java
@@ -10,15 +10,15 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 public class checkIfKickerUpToSpeed extends InstantCommand {
   @SuppressWarnings({ "PMD.UnusedPrivateField", "PMD.SingularField" })
   private final KickerWheel subsystem;
-  public int target;
+  public double target;
 
   /**
    * Checks if the kicker is up to speed
    * <p>Check the variable kickerAtTargetSpeed to check the result of the calculation.
-   * @param subsystem  The subsystem used by this command. (Kicker)
-   * @param target What the target value for the shooter speed is
+   * @param m_subsystem  The subsystem used by this command. (Kicker)
+   * @param m_target What the target value for the shooter speed is
    */
-  public checkIfKickerUpToSpeed(KickerWheel m_subsystem, int m_target) {
+  public checkIfKickerUpToSpeed(KickerWheel m_subsystem, double m_target) {
     subsystem = m_subsystem;
     target = m_target;
     // Use addRequirements() here to declare subsystem dependencies.
